+++ CHANGELOG.md
@@ -6,6 +6,10 @@
   * This calculates the diff for the counter values of pg_stat_statements on the client (i.e. collector),
     instead of the server for increased accuracy and protection against out-of-order processing
 * Introduce "opts" to the snapshot, for indicating which options were chosen
+* Never open more than 1 connection to the same database
+  * This covers edge cases like sending a lot of SIGHUP signals
+  * In case we detect more than 1 connection we error out and exit, to avoid
+    clogging the database
 
 
 ## 0.9.0rc4    2016-04-03